fix: 거래 내역 조회 에러 수정

+++ src/main/java/com/bank/service/AccountServiceImpl.java
@@ -145,7 +145,7 @@ public class AccountServiceImpl implements AccountService {
         transactionMapper.insertTransaction(transaction);
     }
 
-    public List<TransactionDTO> getTransactionsByAccountId(Long accountId) {
+    public List<TransactionDTO> getTransactionsByAccountId(Integer accountId) {
         return transactionMapper.getTransactionsByAccountId(accountId);
     }
